requirements.txt
aio-tg-bot==0.0.9
aiogram==2.14.3
aiohttp==3.7.4.post0
anyio==3.3.1
apirone-api==0.0.1
APScheduler==3.7.0
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bit==0.7.2
bit-help==0.0.8
bitcoin==1.1.39
bitcoinfees==0.0.3
bitcoinlib==0.6.0
blinker==1.4
blockcypher==1.0.91
bs4==0.0.1
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cheroot==8.5.2
CherryPy==18.6.1
coinbase==2.1.0
coincurve==15.0.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==3.4.7
cryptonator==0.0.4
decorator==5.1.0
deprecation==2.1.0
distlib==0.3.2
dnspython==2.1.0
ecdsa==0.16.0
email-validator==1.1.3
emoji==1.4.2
filelock==3.0.12
future==0.18.2
gevent==21.8.0
greenlet==1.1.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httpcore==0.13.6
httpx==0.19.0
hyperframe==6.0.1
idna==3.2
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.3.0
jaraco.text==3.5.1
Jinja2==3.0.1
kaitaistruct==0.9
loguru==0.5.3
lxml==4.6.3
MarkupSafe==2.0.1
more-itertools==8.9.0
multidict==5.1.0
myqiwi==0.2.5
numpy==1.19.5
packaging==21.0
peewee==3.14.4
phonenumbers==8.12.31
Pillow==8.3.1
platformdirs==2.3.0
portend==2.7.1
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.10.1
pydivert==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyQiwiP2P==1.0.1
PySocks==1.7.1
python-capitalist==1.3.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
qrcode==7.3
random-data==0.1.7
random-user-agent==1.0.1
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
scrypt==0.8.18
selenium==3.141.0
selenium-wire==4.4.1
SimpleQIWI==0.7
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
Telethon==1.23.0
tempora==4.1.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
uuid==1.30
validators==0.18.2
virtualenv==20.7.2
virtualenvwrapper-win==1.2.6
webdriver-manager==3.4.2
win32-setctime==1.0.3
wrapcache==1.0.8
wsproto==1.0.0
yandex-money-sdk==0.1.3
yandex.money==0.0.1
yarl==1.6.3
zc.lockfile==2.0
zope.event==4.5.0
zope.interface==5.4.0
